Separate query name resolution from type checking
Summary:
This makes things a lot cleaner: before we were deferring the
resolution of names in queries until typechecking, while the rest of
the schema was resolved earlier. Now we do all the name resolution in
one pass. This will enable the next stage, computing the hash of the
definition of a predicate.

I also cleaned up the resolution pass some more, and addressed feedback from earlier: got rid of `unqualify`.

As part of this I tried to fix the way enums are resolved, but it
turns out to be hard because we currently use the type information
from the context to resolve them (e.g. you write `E : T` to refer to
enum E from type T) and the existing schema relies on that in multiple
places. I'm going to need to do a careful migration to fix it, so
deferring that for later.

Reviewed By: lazamar

Differential Revision: D36675383

fbshipit-source-id: 3c8785c92aafa58d23a56576b0debc76a77909bf