Add database Section type.
Summary:
We want to support scoped seeks in `FactGenerators`.
We will implement it by restricting the range of fact ids in queries to a section of the database.
This diff implements the `Section` type which enforces the range restriction.
Then a `Snapshot` is just a `Section` without the lower bound restriction.

Reviewed By: simonmar

Differential Revision: D36778823

fbshipit-source-id: ed294d49f6f4527be689dfd3637310b7dcbca711