Fix :index lsif command
Summary:
the LSIF indexer requires an lsif file to index, rather than a repo.
This confuses things between regression tests, where we want to specify a particular lsif file but run many quieries in $root,
while in the shell we just want to specify the file name.

Fix it by makign the --input argument optional. if not present we presume indexerRoot is the lsif file.

Reviewed By: pepeiborra

Differential Revision: D36963167

fbshipit-source-id: 2e038d7e916038c52313533a76099db7b27ba588