Hash predicate definitions to produce universal PredicateIds
Summary:
This is the next stage of eliminating schema versions.

Introducing new types `PredicateId` and `TypeId`, which are global identifiers for predicates and types respectively, calculated by hashing their definitions. The idea is that by doing everything in terms of these global identifiers, we'll be able to have multiple versions of the schema around at the same time without getting confused.

There's a new phase in schema processing `ComputeIds`, where we
* take all the resolved schema definitions and compute the dependency graph
* in dependency order, substitute PredicateRef/TypeRef with PredicateId/TypeId and then hash the definition
* with some special treatment for cycles (see comments in the code)

This is all nicely self-contained. After that we can mostly use `PredicateId` and `TypeId` everywhere.

I've removed the `--db-schema-override` option because it was getting too hard to support the previous behaviour. For now we behave as if the option is always on (but validation means that we never encounter any mismatched predicates in production anyway, so this only makes a difference for testing anyway).

Reviewed By: donsbot

Differential Revision: D36779133

fbshipit-source-id: 561eccb401829ca386c727b2bb2324f0d844cd23