Assign fresh Pids to predicates not in the stored schema
Summary:
This is a step towards allowing the global schema to be different from
the stored schema. Where the two schemas overlap they will get the
same Pids, but if the global schema differs from the stored schema,
the predicates of the global schema will get different Pids.

We won't allow this to happen (yet) by continuing to enforce schema
validation.

Reviewed By: lazamar

Differential Revision: D37033749

fbshipit-source-id: 0dd1c93ec6e896d72cd07bcbc1c03b1cdfc9ffa6