Add regression tests for duplicate entities in same file
Summary:
Example of how the entities get mapped in a file that has things with the same name.
E.g. a decl and a defn for 'foo:foo()', we want some examples that will trigger
CxxResolveDeclarationToEntity and friends, and show what is fed into the glass symbol id encoder.

Regenerate with:

> buck run //glean/lang/codemarkup/tests/clang:tests -- --replace glean/lang/codemarkup/tests/clang

Reviewed By: simonmar

Differential Revision: D37133864

fbshipit-source-id: b81cbc3cc475636def7b52171d368f386df9e741