Use a glass-specific set of test C++ files for glass regression tests
Summary:
We want to test slightly different behaviors than the main glean tests.
E.g. handling of decls and defs. This starts down that path

Example shows a collision in the symbol test/cpp/foo/foo, where the
searchEntity result from Glass will be the first of either the decl or the
definition of foo(), depending on order in the file (its ambiguous). Both
are entities, ad codemarkup.cxx.CxxResolveDeclarationToEntity

Reviewed By: simonmar

Differential Revision: D37133314

fbshipit-source-id: d8695ecb60b2394801df07e6bb19f070a4a1942f