Sharding disambiguation
Summary:
Address the concerns in [this comment](https://www.internalfb.com/diff/D37102090?dst_version_fbid=552840562882403&transaction_fbid=1198199747664040) by simonmar:
> We'll need to fix the confusion with our existing APIs that mention shards,

Currently the Glean query server is not sharded / does not support sharding, but it publishes repo hashes as shards to SMC, so that clients can request a node guaranteed to have restored a specific db.

This Diff renames all? the top level bindings and module names to clarify that these are "db shards", and adds comments to explain.

Reviewed By: simonmar

Differential Revision: D37110255

fbshipit-source-id: 8e5375da8875c3fa767aba2adceb9ab4111875bf