Pass schema_id to queries in Haskell clients
Summary:
This does the necessary plumbing to get the `schema_id` at compile
time and pass it to the server at runtime.

No behaviour change yet; the schema_id is not actually used.

I tried a few ways to do this and finally settled on passing it into
`withRemoteBackend` and friends at the call site. The rationale is

* We don't want to pass it with every query, obviously
* We do want to make it visible when a client is pinning the schema
* Some clients explicitly don't want to pin the schema (the shell, the
  CLI tool) and these should not be made to depend on the schema.

So making the choice of whether to depend on the schema in the client
seemed like the right thing, even though it means a small change to
every client.

Reviewed By: donsbot

Differential Revision: D37143728

fbshipit-source-id: e8a49b533f44f3c752cd2fe784c4be1ed1ebafca