Pass database boundaries to bytecode generation
Summary:
To perform restricted range seeks we will need to pass the database boundaries all the way down to bytecode generation.

To avoid breaking the `Stacked` abstraction by exposing its constructors the type `Boundaries` is introduced. It contains the id ranges of the stacked db we are working on and the range that includes all other dbs in the base.

This diff only pipes information. It doesn't change any functionality.

Reviewed By: simonmar

Differential Revision: D36975468

fbshipit-source-id: 19b23fddf46966960fbd47498890cc650930c4b6