Add tests for derived cxx1.FunctionDeclAttribute
Summary:
This is performance sensitive, so check it works and check perf is ok.

Float the attributes/ test out to a top level clang regression test. add a derive step. write a query for the direct predicate, the derived predicate and the cxx.codemarkup entity wrapper.

N.B. I couldn't write a test for codemarkup.EntityAnnotations (functional query) as it segfaults(!) (see next diff)

Reviewed By: simonmar

Differential Revision: D37199638

fbshipit-source-id: a44f2696742be871d9c9156609a4812c10264664