Redefining derived facts should be an error
Summary:
When a query for a functional derived predicate produced facts with
the same key but different values, `define()` would return
`Id::invalid()` but we weren't checking for that. In a debug build it
would lead to an assertion failure (see D37271612) but in a non-debug
build it silently did ... something, I'm not sure what exactly.

Reviewed By: pepeiborra

Differential Revision: D37278501

fbshipit-source-id: 78586828c257fdbdc44e184371b83eb4e9d8a0c4