Take 2: assign fresh Pids to predicates not in the stored schema
Summary:
The handling of Pids in the previous version was wrong in conjunction
with stacked DBs and schema thinning.

Summary from the original diff:

> This is a step towards allowing the global schema to be different
from the stored schema. Where the two schemas overlap they will get the
same Pids, but if the global schema differs from the stored schema,
the predicates of the global schema will get different Pids.
We won't allow this to happen (yet) by continuing to enforce schema
validation.

New in this diff:

* Removed schema thinning. It's not necessary any more (the schema
  thinning test still passes, because we distinguish predicates by
  their hash now.)

* Don't typecheck predicates that we've seen before. This will avoid
  doing any extra work now that we're not thinning the schema.

* Added a test for the previously broken case

Reviewed By: lazamar

Differential Revision: D37208533

fbshipit-source-id: b914eead6338641ace289d04d638af01bf6da120