Fix performance regression in codemarkup.pp.PPEntityLocation
Summary:
Around June 2, there was a regression on the performance of the documentSymbols*() calls to Glass, for C++ of between 2-3x. Details in here: https://fb.workplace.com/groups/csi.wtf/posts/5205339992875958

https://pxl.cl/26pgt

Tracked down to regression in performanec of the query that builds C pre-processor xref symbols in a file, PpFileEntityTraceXRefLocations, and specifically in the entity location builder, PPEntityLocation. Introduced in the check in D36767904 (https://github.com/facebookincubator/Glean/commit/e37f6dc2307d7eb5f909f5d907a11ddbef7043d7)

pp1.Include facts are xrefs to src.Files. the PPEntity location lookup code was
accidentally scanning all uses of a #include, when trying to find the file name of #include target.

```
      pp1.Include { file = File }; # don't need to do this, its a xref scan
      { include_ = File } = Entity;
      src.File Name = File;
      RangeSpan = codemarkup.types.RangeSpan { span = { 0, 0 } }
```

The fix is to avoid doing the full scan of pp1.Include. Just check existence. It's sligthly subtle as we don't have a canonical entity predicate for include_ entities, just their uses. So we need to probe for existence carefully to avoid scanning all uses.

Reviewed By: simonmar

Differential Revision: D37233605

fbshipit-source-id: 20657eeebb4c81604ede644d3c22455be85c15d8