Extend glass:searchRelated with optional symbol descriptions
Summary:
Identical to the detailed results for searchByName, sometimes we want to know all the attributes of the results we get back from searchRelated. This will be used by CodeHub to do 'related' symbol navigation.

E.g. if we list all members of a namespace for API docs, we also need to know their kind, location, sym id, ...

The describe output is often duplicated in relational search, so we return a map keyed by symbol id for descriptions.

Relations that are tested: for C++ and Hack:
- parent `contains` child
- parent `extended by` child
- child `extends` parent

Note we don't yet have in the codemarkup.angle side:
- child `contained by` parent

We want to be very careful to get Haxl parallelism magic in the mapM describe.
The keys to the symbol id map are also strings to avoid double encoding.

Reviewed By: simonmar

Differential Revision: D37372800

fbshipit-source-id: f4298c65b99a569c96b5703de6f65c707f8f977f