Move schemaEnv calculation to ComputeIds
Summary:
This is mainly a refactoring, to move some work from `mkDbSchema` to `computeIds`, because this makes it easier to do this per-schema-instance in the following diffs.

This diff also changes from using `Hash` to identify schema instances internally to using `SchemaId`. This ends up being simpler and will make it easier to write tests later.

Reviewed By: donsbot

Differential Revision: D37378217

fbshipit-source-id: 107efb38ed7f406120d98cbb3dc2815f0291235c