Don't default to using glean.schema_version for queries
Summary:
Rationale:
* Clients should be setting schema_version in queries
anyway. Virtually all of our clients do this now.
* In the next diff, I'm getting a bit stricter about what's in
scope. Only the predicates actually reachable from the all.N schema
will be in scope. Together with the glean.schema_version behaviour,
this made some tests fail.
* Defaulting to using the schema from the DB causes some confusing
behaviour, e.g. we recently had some DBs being produced with
`glean.schema_version = 8` and some predicates were not available in
the shell.
* There's no good way to override the default behaviour to say "I want
the latest". The alternative to this is to add an explicit way
to get the latest version, which seems worse.

Reviewed By: donsbot

Differential Revision: D37413967

fbshipit-source-id: 6fe18962abaf606d24b6211a9f9b530ac0b0e01d