fix a laziness bug
Summary:
We were taking the length of the DB tar archive bytes stream, causing it to get eagerly materialized and leading to an OOM crash-loop.

This is a minimal fix that removes the `LBS.length` call and instead reuses the `dataSize` which is already available once the stream has been uploaded.

Reviewed By: simonmar

Differential Revision: D37485497

fbshipit-source-id: 3c9fd5549fff90e1883475b8041568f9853d5966