Bulk fetch locations when doing glass.SearchRelated
Summary:
Get rid of extra pass of fetching symbol locations in SearchRelated results. We always need the locations, so use the combined searchRelated query that grabs them for us in one pass.
Cleaner and more efficient. Removes a ton of duplication and the relation type and direction are pushed to Glean, so we just dispatch in Angle to difference searches.

Reviewed By: nhawkes

Differential Revision: D37449881

fbshipit-source-id: 071792be92c1469c6815872ca2d40bb7a57095f0