New validity check for schema changes
Summary:
When adding a new schema instance, predicates in the new schema must
be compatible with the same predicates in the older instances of the
schema, where "same" means "same name and version".

We'll test this in `glean/schema/sync-cf`, but it will only require
looking at the SchemaIndex in configerator. There will be a validation
that we have to perform on the servers, but it's only to check that
the servers can parse the schema.

Reviewed By: pepeiborra

Differential Revision: D37514017

fbshipit-source-id: d554086c84edc6836203e33ae12474543dea0244