Store all function results via pointers
Summary:
The `std::function`s we called from bytecode produced their first
result via `return` and the other by writing through pointers that were passed
in. The diff removes the special treatment of the first result - it is now also
written through a pointer and the functions are always `void`. This isn't an
interesting change in itself but will allow nicer things going forward.

The diff also fixes some infelicities in the types of `std::function`s pass in
in `query.cpp`.

This doesn't need a bytecode version bump.

Reviewed By: simonmar

Differential Revision: D37546806

fbshipit-source-id: 1b32e65f24a8e9321b5716a4eb351c83224bf1b2