Expand query in incremental derivation
Summary:
This diff contains the transformation to make a derived predicate query yield only results which make use of new information.

Once this is merged the server will automatically detect incremental derivations and trigger the transformation without any prompt from the client.

As per the explanation in `Note [Incremental derivation - Implementation]` in the code below, this is `O(n^2)` where n is the number of `FactGenerator`s in the longest branch of the query.

I believe the complexity of this transformation shouldn't be a problem because.
- It will be applied only during the derivation step of the indexing process
- It will be run only once per derived predicate
- The continuations don't travel to the client
- Derivations usually have at most a dozen or so `FactGenerator`s in the longest branch.

-----
## Could it be linear?

Yes. And this might become important when we use some version of this transformation to handle recursive queries.

First we note that although the final query bytecode is `O(n^2)` the time complexity of the transformation algorithm and the size of the tree of `CgStatement` are both `O(n)` because we share many intermediate nodes.
```
P A; Q B; R C; S D
```
becomes

{F746875556}

If we can share the compiled code for the common paths we can bring the bytecode to `O(n)` as well.

Reviewed By: simonmar

Differential Revision: D37414045

fbshipit-source-id: c7d685e1ee685fb0e1065908fbf24b2833c5082a