Fix a bad error message
Summary:
A call to `error` here doesn't kill the server, only the Janitor thread, which will be restarted 2m later by the Periodic wrapper and presumably fail again until the underlying problem is solved.

The problem is that we don't download dbs in dependency order, so it's perfectly possible that a dependency is missing transitorily. Until then, we cannot calculate the shard id. I think this will effectively enforce a dependency order for downloading dbs :)

Reviewed By: simonmar

Differential Revision: D37613791

fbshipit-source-id: bd3afac9dbb9ddd871da1653214cb8545ce753ff