Wait for shard assignment
Summary:
Reviewing the logs, I just noticed that the Janitor runs before the SM client has received the shard assignment, and schedules all the local dbs for deletion. This will generate a lot of DB churn.

Ideally we would wait until the shard assignment is complete, but I don't see any way to do that. So instead, wait for a fixed amount of time. Looking at the server logs it takes around 30s for the shard assignment to complete, so let's wait for 45s to be sure.

Reviewed By: simonmar

Differential Revision: D37643689

fbshipit-source-id: 6a7a01184647851bafbd523daf832299374171a4