Change how suspending works
Summary:
Previously, we were calling a function which recorded the state and then returning from the subroutine. Now, we update the Activation such that then next call to `execute` will resume the subroutine and return a status code from the evaluator which says whether the subroutine has been suspended. This means that we decouple serialisation/deserialisation of suspended subroutines (subsequent diffs) and can also be easily tweaked to support subrotuine calls from within the bytecode should we need that.

Note that this does *not* require a bytecode version bump as the implementation is both forwards and backwards compatible. This required leaving in the register which was used for passing in the `saveState` function for now. It is unused and will be removed later but I'd rather do all breaking changes in one go.

Reviewed By: simonmar

Differential Revision: D37602267

fbshipit-source-id: bf35f4566c80a71e1208ebd89fa89786fabe173e