Introduce subroutine activation records
Summary:
This encapsulates some of the functionality needed to prepare and execute a `Subroutine` in the new `Activation` type. This is both cleaner and will allow more nice things further down the road. A slight benefit is that we can now write arguments directly into the frame rather than creating an intermediary `vector`.

This would be very easy to extend to allow subroutine calls from within bytecode if we have a use case for that.

Reviewed By: simonmar

Differential Revision: D37555822

fbshipit-source-id: bf54bdb12b3ae61f11316bb9dcfb325c879b7129