extract repoTimeOpt as parser
Summary: Personally for me, the code of the function is too long and hard to follow. Also I get used to have all parameters specified somewhere at one place at the top of the file, so it is easier to edit and see. So I have started just extracting all parameters into separate functions at the top of the file

Reviewed By: simonmar

Differential Revision: D37682803

fbshipit-source-id: c634764596add6816f17a21610d20d67b8dd6dd5