Unset the expiring bit for DBs we want to keep
Summary:
We want to handle the following scenario:

1. Server starts up and finds a db A in the local storage
2. Janitor kicks in with an incomplete shard assignment that doesn't include A, so A gets marked for expiry
3. In the next Janitor run the shard assignment is extended with A, which gets added to the `keep` set

Before this Diff, 3 would be followed by:
4. After roughly 30m (the retention expire delay for glean.query.prod) the DB gets deleted. This is not what we want!

After this Diff, 4 will not happen. Moreover, we can drop the sleep when the shard assignment is empty, since there is basically no penalty for starting with an empty or incomplete shard assignment (assuming a long enough expire delay).

Reviewed By: malanka

Differential Revision: D37715592

fbshipit-source-id: 4ef9384a5e0cc12befab5424dd907bc4306bfc39