Support updating the schema in a stacked DB
Summary:
We sometimes want to be able to use stacked DBs for testing derived
predicates, but doing that safely is a bit tricky. The workflow will
be

```
glean create --stacked <base> --repo <repo> --update-schema-for-stacked
```

To use the current schema in the stacked DB, instead of the schema
from the base DB. This will fail if the current schema is
incompatible; you can only add or remove predicates.

Reviewed By: pepeiborra

Differential Revision: D37648917

fbshipit-source-id: 5f1f6a7f60319de350426e1c49bb1534fd5dcbdc