Fix a bug in schema processing
Summary:
In typecheckSchema we were calling `checkStored` on all the predicates
in the environment, but using the type environment from the current
schema only. So if any types were removed or changed, we would get an
error.

The fix is to not do `checkStored` on predicates that were already in the
`tcEnv`, because those have been checked already.

Reviewed By: pepeiborra

Differential Revision: D37786898

fbshipit-source-id: 50af80439b07a03231fd066cab619a917f555b5a