Implement entity language filters in search on Glean side
Summary:
This extends/replaces the generic searchByName with kind filters, with an additional language filter.

Meat is in:
> codemarkup.search.SearchByName
> code.EntityLanguage

And as Simon suggested, push the filter into the case analysis. This is actually a good win.
```
www.hack> search.code.SearchByNameAndLanguage { name = "Glean", language = Hack }
1 results, 4 facts, 2.81ms, 3482224 bytes, 7640 compiled bytes
Facts searched:
                             hack.Name.6 : 1
                 hack.ClassDeclaration.6 : 1
                            hack.QName.6 : 1
www.hack>
```

The language enum is a direct case analysis against the code.Entity tags.

Filtering is cheap and takes place after finding entities. In the common search case the language filter is wild, though codehub supports sets of lang:foo queries which we translate into A|B disjunctions on the search results.

This will get a lot more useful once Thrift, Buck and PreProcessor syms start showing up in searchByName in fbsource.

The plan - on the glass side - is to pass specific language filters through where the user has specified them, avoiding the need for filters in the glass client. We'll then generalize glass.thrift:searchByName to take a set[Language] filters, which will let us delete a ton of codehub client code for probing for languages.

Reviewed By: pepeiborra

Differential Revision: D37987282

fbshipit-source-id: ea415662f42aeb246318e2a464b7f9c030ad9dc2