Handle UnknownException from expireDatabase
Summary:
The Janitor can call `expireDatabase` with a no longer `Local` DB, if the DB gets deleted in between the Janitor reading the Catalog and the call to `expireDatabase`, which will result in a legit `UnknownDatabase` exception.

We could fix this by making the Janitor transactional, but I am unsure whether the additional code complexity is worth the benefit.

The simpler fix is to ignore the `UnknownDatabase` exception since it is benign, as the DB is already deleted.

Reviewed By: simonmar

Differential Revision: D37994132

fbshipit-source-id: 32869188a9e0f2c7f15737ae50258becfe6b07c2