Track the pre-transformed type in query continuations
Summary:
We had a query failure when doing a schema migration caused because the transformed type could not be reversed back to the original type. This avoids the problem by using the pre-transformed type from the typechecker.

This is not only simpler but it means we can give a more useful type in the `resType` field, which will improve our logging.

Reviewed By: lazamar

Differential Revision: D38038537

fbshipit-source-id: 82b4fffe86ee75e96df0412834a8102ee574667a