Implement language filters on search results on Glean side
Summary:
This moves the optional entity language filter from the Glass to the Glean end of the wire.

This should be slightly more efficient in the rare case we filter repos with >1 language entity type in them (e.g. fbsource soon). It also sets us up to generalize search to take a keyset of languages, instead of just 0 or 1 language filter.

*Don't land until codemarkup.search.SearchByName is on the prod query servers*

Reviewed By: pepeiborra

Differential Revision: D37987728

fbshipit-source-id: 4f7d998a2a95577866daa8686a70dcc370c83373