exclude/include from file
Summary:
Currently, we pass units to be included/excluded in an incremental db from the CLI as a sequence. This doesn't scale when passing large number of units.

This diffs adds two options to include and exclude unit from a file instead of by passing the units on the CLI: `--include-file` and `--exclude-file`.

Reviewed By: donsbot

Differential Revision: D37969392

fbshipit-source-id: 43584a2830be60defa0c8a558eb68eaa4338bcb0