Treat ObjectiveC language filters as Cpp for search purposes
Summary:
We don't have a code.Entity tag for objective C entities, we treat them all as Cpp.
This is probably not the right thing to do.

For now, if the user specifies ObjectiveC as a langauge filter, treat it as Cpp so we can find "fbsource", which really does contain ObjC code.

The CodeHub client will remap the entity language by filepath so its not terrible.

Reviewed By: pepeiborra

Differential Revision: D38266026

fbshipit-source-id: ccfd5e895c796ce222bd4a92d7174fad17851763