Open only the assigned most recent DBs
Summary:
The Janitor unconditionally opens the most recent DB of each repo found in the local disk. This behaviour assumes:
1. that the most recent DB is always in the shard assignment, which is false when a non trivial sharding policy is used,
2. that the most recent DB is not currenly being deleted, which leads to non-fatal but noisy UnknownDatabase exceptions in the logs

The motivation for this trick (D16358893) is to avoid lag spikes when a DB is opened for the first time.

This Diff scopes the behaviour to only DBs in the shard assignment which are not scheduled for deletion, hopefully addressing both issues.

Reviewed By: donsbot

Differential Revision: D38150397

fbshipit-source-id: d3cb19a47322333eaed9b1991e1eae1a0d43e0ff