Work around: Be more tolerant of non-qnames when parsing
Summary:
The current serde parser processes any json with label "value" as a possible type, looking for qnames. However, the rust save-analysis output uses "value" for types and for comments/docs. So we end up trying (and failing) to parse some comments as qnames.

Instead of dying when this happens, skip the token, so we can continue indexing at least.

The real fix is to update the JSON mapping for save-analysis to ignore/record the hover/comment content separately to types.

Reviewed By: pcwalton

Differential Revision: D38306196

fbshipit-source-id: 0a11da017e696620b7cebc47b6f62d1af2153f84