Fix dbAge counters
Summary:
I noticed a discrepancy between the `fbsource.fbobjc.buck.targets` db age detector:

{https://pxl.cl/2970w}

And the actual db age (10h) reported by the `glean` command:
```
fbsource.fbobjc.buck.targets/ae2970d8402c4bdf5d65a5e975959e6acff8bea6 (complete)
  Created: 2022-08-02 22:11:11 UTC (11 hours ago)
  Completed: 2022-08-02 22:43:06 UTC (10 hours ago)
  Size: 289.48 MB
  Write server: glean.write.frc
  Replication: 19/36 of query servers
```

It turns out that db age reporting was broken by sharding. For every db name R, the glean server publishes a counter `R.age` with the minimum of the ages of the R db instances it carries. Before sharding, this was indeed the age of the youngest R db. But with sharding, the set of R db instances that a server sees is partial and may not include the newest instance.

The end result is that there is a lot of variance in the ages reported for a db-name R, and since our monitoring config does not specify a reduction strategy I can only guess it's picking a random one:

{https://pxl.cl/29710}

The fix is to only report db age if the server sees the newest db instance. We already use this condition for other code paths, like auto-opening the most recent db instances.

Reviewed By: malanka

Differential Revision: D38385201

fbshipit-source-id: fafb70e0df47180347fbc7c87049265db789f9c2