add a 'most recent' sharding policy to increase capacity
Summary:
Introduce a new sharding policy that places most recent DB instances in a designated shard. Our SM config replicates this shard to all server; this is desirable since these DB instances assumedly carry >90% of all traffic.

This can be landed and deployed now, but the new sharding policy is not quite compatible with the current SM config as the published counters now carry the domain id. We can change the current config to make it compatible as follows:

1. D38423294 Update the old sharding policy to publish counters with and without domain id (and deploy).
3. D38423319 Update the config with the new mostRecent domain, using the new counters. This is fine as it is compatible with the updated old sharding policy.
4. D38426272 Switch to the new sharding policy

Reviewed By: lazamar

Differential Revision: D38423099

fbshipit-source-id: 6ee3d1b3bbfb6b8607569cdf5c242e040898c28c