Wire up namespace search to searchSymbol
Summary:
When namespaceSearch is set parse the query string as tokens and issue a codemarkup.search.SearchByScope query.

If the string parsers as a tokenized string literal (eg ::foo) we will do a search for 'foo' as well.

For now I'm not supporting `scope::_` wild card style, until we get the indexing right. So the following is expected to work ok:

```
::a
a::b
a::b::c
::a::b
```

But not anthing ending with a scope wild card `a::b::`

Reviewed By: nhawkes

Differential Revision: D38415386

fbshipit-source-id: 8ce9018f45a7d78b3bdb35e6b0773a1c0c80a8cc