Say how an instruction affects control flow when issuing it
Summary:
Previously, we were recovering that information from the instruction itself via the codegen'ed `insnControl`. Now, we have a separate `issue` function for each kind of instruction (fall through, conditional jump, unconditional jump). Care has to be taken to use the right `issue` but we codegen the functions which actually call `issue` anyway so clients should never have to care about it. A future refactoring might make this explicit in the module structure.

This doesn't seem to noticeably improve performance, sadly, but it makes `issue` lazy in the issued instruction where it was strict in it previously which is necessary for some changes in future diffs.

Reviewed By: lazamar

Differential Revision: D38454951

fbshipit-source-id: 268231a38daa6a89e615c3b9373f3bbf130ffadc