Make Insn monomorphic
Summary: The instruction type `Insn` (which we codegen) used to be polymorphic in the register and label types. Now that we only have one instance of each, let's make it monomorphic. This hopefully leads to a more compact representation and might enable further improvements in the future (like getting rid of `Insn` entirely).

Reviewed By: lazamar

Differential Revision: D38454949

fbshipit-source-id: 11781542f5c61bd265e6440138e42475b24af21d