Use one word to represent registers during codegen
Summary: During codegen, register have a "segment" and an index within the segment. This diff squashes the two into one word, using the two high bits for the segment. This doesn't really make the code a lot worse but enables some goodies in future diffs (in particular, a monomorphic `Insn`).

Reviewed By: lazamar

Differential Revision: D38301925

fbshipit-source-id: 8458fe890cc9dadc933f5e4cead1ffe826a7444b