Prune fruitless query branches
Summary:
When we open the db we know what predicates contain no facts.
We can go through derivation queries and remove all branches that depend on facts from those predicates.

This means that a `codemarkup` query for example, would be specialised to a specific db and the inclusion of numerous other languages in the derivation would not influence its compilation time.

Reviewed By: simonmar

Differential Revision: D38312646

fbshipit-source-id: be6255d8d194a3e41faceafdbc4c526a66b96571