Fix tests
Summary:
Looks like we are having quite a few predicates.

Finding out was that this was the issue was not very fun so it is worth writing a test to check this.

What is going on?
The pruning script wasn't finding a predicate which should be in the db in the status Map for the Memory storage method because the map didn't have the predicate's Pid.

The memory storage method is only used for testing.

Reviewed By: malanka

Differential Revision: D38631020

fbshipit-source-id: 6dc44d3cd74d68c7e7438e56c015fdb70a73f6e6