Add a #include xref resolver
Summary:
This is another use case where the full symbol data for a file isn't needed. Instead, we just need to know the filepaths the xref (#include xrefs only) refer to.

Implement a thrift method for recursively pre-fetching the xref target file paths. Currently this is 100% special for C++ and #includes but in principle we couuld pass a symbol kind filter and do it generically ,if there's other kinds of targets we want to pre-fetch.

It's a semantic-aware cache warmer.

Testing some perf on  fbcode/biggrep/BigGrepClient.cpp, which has nice fan out of #includes at depth 2 we get a p90 of 27.8ms for 364 results. Not bad!

https://pxl.cl/29JCD

Reviewed By: pepeiborra, williamtwilson

Differential Revision: D38610949

fbshipit-source-id: 60b8188f6e42da2f854d7ed63a0b4eadcb28c9cb