Avoid opening DBs which are deleting
Summary:
Opening a DB that is scheduled for deletion fails with an UnknownDatabase
exception. The exception gets handled and logged, which pollutes the logs. Moreover,
the deletion attempt fails and gets rescheduled. It succeeds in production because
the ordering of the Janitor side effects and the Janitor period align, but it confuses
tests where the Janitor period is arbitrary

Reviewed By: malanka

Differential Revision: D38567984

fbshipit-source-id: 886f04ecf4c4d95f91be50bc52e298a63a004ee0