Fix sorting of DBs
Summary:
This fixes the bug that made the db age counters disappear.

However, there's one more issue left: the query for computing the set of most recent DBs is not looking at `Cloud` DBs. So if there are any newer DBs that don't belong to the shard assignment, the Janitor won't see them and publish an incorrect value for the db age counter. The test failure reflects this problem, and the next Diff attempts a fix.

Reviewed By: malanka

Differential Revision: D38567981

fbshipit-source-id: bab7a9e2feef7f26362f0e34cafb5355be6e9c37