revisit symbol-ids
Summary:
Currently, this is what the three types of symbol ids look like

1. Definition `fbsource/buck/fbcode/calibra/macros/defs.bzl/python_library`
2. Source file `fbsource/buck/fbcode/libra/wallet_media_storage/services/mock/gk/mock_gk_service.py`
3. Locator/target `fbsource/buck/fbcode/calibra%2Flib%2Fauth/auth_context_lib `

After this diff, it becomes

1. Definition `fbsource/buck/d/fbcode/calibra/macros/defs.bzl/python_library`
2. Source file `fbsource/buck/f/fbcode/libra/wallet_media_storage/services/mock/gk/mock_gk_service.py`
3. Locator/Target `fbsource/buck/t/fbcode/PATH/tarcalibra/lib/auth/auth_context_lib `

For TARGET, we need to identify the subdir from the rest of the path, e.g. fbcode/X/Y and arvr/X/Y need to be decoded respectively as `{subdir= {just = fbcode}, path="X/Y"}` and `{subdir = None, path ="arvr/X/Y"}`. To avoid ambiguity, I added an extra tag "PATH" to mark the separation between the subdir and the path.

This isn't ideal because PATH could be an exist dir name and looks a bit ugly in the middle of the symbol id. We could maybe use an empty string instead.

Critical signal sigma_package_builder-haxlsh seems unrelated to this diff (sandcastle job has been failing for many other diffs).

Reviewed By: donsbot

Differential Revision: D38435666

fbshipit-source-id: 317796e7e7e3addf839b42720ad770c992babc99