Let functions called from bytecode decode their arguments
Summary: The main point of this diff is to make the implementation of the `Call_M_N` instructions exactly the same. This will allow us to only have one call insn later (D38031725). To do this, we push the decoding of arguments into the syscalls (new name for C++ functions called from bytecode) and introduce combinators for defining those.

Reviewed By: simonmar

Differential Revision: D37702210

fbshipit-source-id: 884be528afe428bab8d1dcc06bfb1eb8e4ed9942