Don't use std::function in syscalls
Summary: I've always disliked this since `std::function` allocates and we had to be careful about sharing them between invocations. Now, we pass in a `context` object which is then passed to syscall handlers and the combinators for defining those handlers are modified to work with methods of the `context` object. This seems much nicer to me and guarantees no allocation at runrime.

Reviewed By: simonmar

Differential Revision: D37702211

fbshipit-source-id: c29cde5728d433fbadf1a64cbd0df8cf412fb8ee