Don't dynamically allocate for small binary::Output
Summary:
This implements the standard small buffer optimisation where we use the memory occupied by the object itself to store the actual data if it fits. Details are in comments.

This seems to make some benchmarks (`rename-bench`, parts of `query-bench`) about 10% faster for me, although the data is very noisy.

Reviewed By: simonmar

Differential Revision: D38517366

fbshipit-source-id: 4db5b8cee9714d5cc60ac596c0be66484e4cf985