Teach glass how to find function local param entities
Summary:
When the 'local' scope variety was added for C++ local variables, we also needed to update glass to search for such things. Impact: glass should be able to resolve symbol ids for locals now in most cases.

We've been correctly generating symbol ids for the scope, but not the search term to find them. Fix that.

This ensures we don't silently fail to resolve local param symbol ids.

Regen tests with:

>
and
> buck run mode/opt //glean/glass/test/regression:cpp -- --replace glean/glass/test/regression/tests/cpp

Reviewed By: tdupes

Differential Revision: D38527872

fbshipit-source-id: f0a881fe4060ac40de37db7a1b70bb8d93a5a8db