Add a cache to the schema parser
Summary:
Speed up processing the schema index by 3-4x with the current ~20
schema copies.

We want to reduce the cost of having many copies of the schema.
Profiling showed that most of the time spent processing the schema is
in the parser, and since a lot of the schemas we're parsing are
identical we can use a cache to speed this up a lot.

The idea is very simple: just split the input on the "schema" keyword,
and use the hash of the string as a cache key. This should also reduce
the memory requirements, since we only store one copy of each unique
parsed schema.

You could imagine changing the storage layout to store schema
fragments by key so as to avoid storing the same schema twice. Or even
doing it on a finer granularity, say at the level of predicates and
types, but all that is more complex and probably not worth the effort.

Reviewed By: donsbot

Differential Revision: D38657711

fbshipit-source-id: 362c6287739f443017f5fe3bfa599297d80e66a5